Add request limiting middleware to server
This commit introduces a middleware to limit the number of concurrent HTTP requests the server can handle. The limit has been set to a default value of 500 but can be changed in the configuration. This is a measure to prevent potential memory leaks and keep the application robust under high load conditions. The middleware is provided by the `github.com/easonlin404/limit` package.