Replace custom load subset with gopher-lua-libs preload
Due to the code complexity and maintenance overhead, we have decided to remove the custom subset loading of 'gopher-lua-libs'. Instead, we have added 'gopher-lua-libs' to the preload section of 'util.go', 'action.go', 'filter.go', and 'feature.go'. This update streamlines our codebase by using the built-in preload function provided by the 'gopher-lua-libs'.