Add CPU usage monitoring and improve signal handling
CPU usage percentage has been added to the Prometheus metrics monitored by the application. The feature uses the gopsutil library to calculate the current CPU usage and updates the value in a gauge metric. There was also an improvement in the function that handles operating system signals - the individual termination function now stops statistics from being tracked, removing the need for a separate 'done' channel. Last but not least, several Go dependencies were added as well to support these changes.