Refactor Prometheus FunctionDuration labels with task names
A new helper function, `getTaskName`, has been created to reflect both the action name and the script path in the Prometheus FunctionDuration labels. This improves granularity and specificity when measuring and monitoring script performance. In addition, a `getLuaActionName` function is implemented to derive action names based on different types of Lua actions.