Refactor startHTTPServer function to use contextStore
The function startHTTPServer has been refactored to have an additional parameter, the contextStore. As part of this change, a server context was added to the contextStore struct. These changes improve the context management within the HTTP server, making it clearer and easier to track. Moreover, the http server shutdown method is adjusted to use a deadline context to ensure proper shutdown process within a specified time frame.