Add handling for SIGUSR1 signal and restart functionality
A new functionality to listen to and handle SIGUSR1 signals has been added in handleSignals function. This enhancement includes a new function, handleUsr1Signal, to listen for the SIGUSR1 signal and a function to handle server restarts, handleServerRestart. The contextStore structure was modified by removing the sql field and adding the server field to cater this change. In the HTTP server start-up sequence, a condition was added to make sure the HTTP end signal channel is only instantiated if it hasn't already been.