Refactor brute force protection protocol handling
A new processing function `processProtocols()` was added to better handle protocols in LoadableConfig. This move centralizes protocol configuration to the `LoadableConfig.Server` object, removing the need to handle it in the `EnvConfig`. References to the older `EnvConfig.BruteForce` have also been updated to reference the new `LoadableConfig.Server.BruteForceProtocols`. This change results in cleaner, more organized and easier-to-read code across several files.