Introduce Redis IdlePoolSize configuration
There are several changes made to ensure IdlePoolSize configuration for Redis is supported and handled properly in the server program. An IdlePoolSize field has been added in the server configuration. If a negative IdlePoolSize value is set in file.go, it will now be considered as zero. These IdlePoolSize values are applied in util.go when creating new Redis clients. Lastly, an error message was corrected in main.go.