Add monitoring flags to authentication
The code now incorporates monitoring flags within the Authentication struct. These flags indicate if certain monitoring aspects are turned on or off. A new function, haveMonitoringFlag, has been implemented to check the presence of a flag. The flow has been adjusted in several places to accommodate these flags - helping to customize handling of in-memory and cache scenarios.