Refactor Lua backend and move LuaBackendResult struct to lualib
This commit focuses on restructuring the Lua backend-related code. The struct LuaBackendResult, along with its linked functions and methods, have been relocated from 'server/backend/lua.go' to a new file in 'server/lualib/backendresult.go'. Moreover, the LuaBackendResult methods are now initiated when the Lua pool is created. This results in cleaner code and improves the overall modularity of the codebase.

Signed-off-by: Christian Roessner <c@roessner.co>