Refactor error handling in action.go
This revision improves the error handling logic in the action.go file. Specifically, it moves the extraction of an integer from the Lua stack, and the subsequent DebugModule information logging into an else branch to only occur when there's no execution errors. This makes the sequencing and the dependency on the absence of errors much clearer.

Signed-off-by: Christian Roessner <c@roessner.co>