Refactor HTTP request handling and add LDAP functionality to Lua library
This update includes a significant refactor of HTTP request handling in the Lua library, switching from explicit handling of http.Request objects to using the more flexible gin.Context. This allows greater adaptability and extensibility for future developments.

Additionally, this commit introduces new LDAP functionality to the Lua library, allowing scripts to handle LDAP requests and responses directly. This aids in expanding the capabilities of scripts and enabling advanced behaviours.

Signed-off-by: Christian Roessner <c@roessner.co>