Update build process in GitHub Actions workflows
The build process for both 'build-stable.yaml' and 'build-features.yaml' GitHub Actions workflows has been revised. This change adds a step to navigate into the 'server' directory before performing the 'go build' command. It ensures the correct build context for the Go application, improving the reliability of the build process.

Signed-off-by: Christian Roessner <c@roessner.co>