Refactor login and registration process
The commit included several changes to the login and registration process in both back-end and front-end code. On the back-end, it transitioned the registration process to use the login process instead. This was done by removing the original register.html file and adjusting register.go to use login.html file. In addition, the logic for handling successful and unsuccessful authentication results was separated into their own functions for better code readability and maintenance. On the front-end, an option for remembering user details was added to the login.html file. As part of these changes, error handling and redirecting procedures were also updated.

Signed-off-by: Christian Roessner <c@roessner.co>