Refactor login handler and add TOTP login support
This commit involves a significant refactor of the server code related to the login handler. It improves procedural separation and enhances code readability. Notably, this update includes a new script for displaying the TOTP (Time-Based One-Time Password) login page and a script for verification and redirection in another flow depending on TOTP. It's a part of implementing two-factor authentication support for the application.

Signed-off-by: Christian Roessner <c@roessner.co>