Refactor createLanguagePassive function to include destination page
This commit modifies the createLanguagePassive function to use specific destination pages instead of the default login page. The function now takes in "destPage" as an argument and uses that for constructing the `languageLink` instead of defaulting to the "login_page". This change affects all places where the function is used, providing links to the relevant pages based on context.

Signed-off-by: Christian Roessner <c@roessner.co>