Refactor authentication result handling in register.go
Revised code in register.go to distinguish successful and failed 2FA authentication outcomes. Removed an unnecessary blank line and renamed a variable to better reflect its functionality. Adjusted the function processAuthResult, adding authCompleteWithFail as an extra parameter and differentiating actions based on outcome. This helps to handle the 2FA redirection more appropriately based on whether authentication was successful or failed.

Signed-off-by: Christian Roessner <c@roessner.co>