Add optional PoolOnly configuration for LDAP
This commit introduces an optional LDAP configuration setting: PoolOnly. When enabled, it bypasses the creation and close operations of channels and goroutines related to authentication (LDAPAuth). Furthermore, the commit ensures PoolOnly is checked before executing context and worker processes, improving efficiency for applications that do not require the LDAP authentication feature.

Signed-off-by: Christian Roessner <c@roessner.co>