Refactor auth request processing and TOTP secret retrieval
In this commit, we made changes in the authentication mechanism and the TOTP secret retrieval process. Firstly, we simplified our condition in the authentication process where the 'found' variable was removed to reduce unnecessary code complexity. Secondly, we updated the TOTP secret retrieval in LDAP to immediately return the value, improving the readability and efficiency of our code.

Signed-off-by: Christian Roessner <c@roessner.co>