Refactor code for secrets and session management
The configuration of certain secret keys and session management has been refactored. The handling of CSRFSecret, CookieStoreAuthKey, CookieStoreEncKey and PasswordNonce has been moved under specific configurations (Frontend and Redis). Frontend related setup and endpoints are only initialized if the frontend is enabled in the configuration. This makes the server configuration more organized and clearer to manage.

Signed-off-by: Christian Roessner <c@roessner.co>