Add callback service to authentication process
The callback service has been added to the server's authentication process. This addition includes setting up the authentication for the callback service as well as executing the callback in the HTTP and REST layers of the process. The callback service aims to call a generic Lua function. Future development will finalize this functionality.

Signed-off-by: Christian Roessner <c@roessner.co>