Add Lua callback feature to main server code
Introduce the Lua callback feature in the authentication server logic. This update refactored the setupFeaturesAndFilters function to setupLuaScripts, which now also pre-compiles Lua callbacks. Additionally, the logic for executing Lua callbacks was added to the server's REST API and integrating it into configuration handling. Errors encountered are now captured and logged.

Signed-off-by: Christian Roessner <c@roessner.co>