Add RedisHExists function to check field existence in Redis hash
The new function, RedisHExists, has been added to the lualib. It checks whether a particular field exists in a Redis hash stored at a specified key. It interacts with the Redis instance via the ReadHandle, providing an error message or the existence of the field. Also, the function name 'redis_hexists' has been added to global const for ease of use.

Signed-off-by: Christian Roessner <c@roessner.co>