Refactor server setup and add proxy protocol support
The commit refactors the setup process for the IMAP and SMTP servers in our 'go' implementation. It covers arrangements related to using 'proxyproto', and changes the login log message to only print when a new connection is established. For both IMAP and SMTP servers, environment variable parsing and error handling are revised, and default values are added for server address and server name variables, if not provided.

Signed-off-by: Christian Roessner <c@roessner.co>