Update authentication check in monitoring.lua
The authentication check in monitoring.lua is updated to include a consideration for passdb requests. This check is used to determine if a backend server search should be conducted. Moreover, to ensure safe execution, an additional check is added that only allows iterating through backend servers when their number is greater than zero.

Signed-off-by: Christian Roessner <c@roessner.co>