Add function to remove attributes from backend result
A new function 'removeFromBackendResult' is added to the 'filter.go' file which allows for the removal of attributes from the backend result set. It is implemented in the Lua library and integrated into the 'CallFilterLua' function. Modifications were also done to accommodate this change in 'server/core/auth.go', and the relevant constant was added in 'global/const.go'. Minor adjustments in the authentication operation mode and local cache TTL configuration were also done.

Signed-off-by: Christian Roessner <c@roessner.co>