Add version flag and instance info metrics
Two main functionalities were added to the server main file: the addition of a command line version flag that when set, prints the application's version and exits. The other addition is an initializeInstanceInfo function that initializes the InstanceInfo metric with the instance's name and version. Both functionalities have been implemented in the main function of the server.

Signed-off-by: Christian Roessner <c@roessner.co>