Add logging setup and logging options to Lua callback function
A new function, setupLogging, has been created to set up log_format and log_level based on the LoadableConfig.Server.Log settings. This function creates a Lua table and is called in the executeAndHandleError function. Changes are also made in the Lua script to handle this log table and print logs in either JSON or standard format based on the log_format setting. Constants for log levels and formats have also been added to const.go.

Signed-off-by: Christian Roessner <c@roessner.co>