Refactor SMTP client code for sending emails
The refactor consists of the creation of a MailOptions struct to group all parameters needed when sending an email, which helps clarifying code purpose. The RealSMTP struct was renamed to RealSMTPClient to add clarity. Various sections have been updated to use this new struct, resulting in a more maintainable and understandable codebase.

Signed-off-by: Christian Roessner <c@roessner.co>