Update SMTP client authentication and recipients setup
This commit modifies the SMTP client's authentication method to use a server with a specific port if TLS is enabled, startTLS is not, and the port is 465. Also, it involves a slight adjustment in the recipients setup in the smtp.go file, initializing the 'to' slice with zero length, and letting 'append' handle the dynamic sizing.

Signed-off-by: Christian Roessner <c@roessner.co>