Update SMTP client interface and usage
The SMTP client implementation has been updated. This update involves reorganizing the code into the smtp sub-package and modifying how it's used throughout the project. Instead of the old interface, RealSMTPClient in the root of the lualib, the smtp package now exports an interface called Client. This interface is more concise, reducing the complexity and making it easier to use. Along with it, the EmailClient struct is now used throughout the project to send emails.

Signed-off-by: Christian Roessner <c@roessner.co>